Set RequestedTheme immediately after InitializeComponent
The `RequestedTheme` property is now set right after the `InitializeComponent()` call in the `App` class constructor, ensuring the application theme is applied immediately upon initialization.

The redundant setting of `RequestedTheme` in the `MainWindow` initialization has been removed, as it is no longer needed due to the earlier assignment in the constructor.

Signed-off-by: Christoph Hofmann <oss@hoffe.org>